Leopard + 64bits

+++ HotKeyToolKit/Headers/HKKeyMap.h
@@ -258,4 +258,4 @@ typedef enum {
 	@result     Return a carbon modifier.
  */
 HK_EXPORT
-UInt32 HKUtilsConvertModifier(UInt32 modifier, HKModifierFormat input, HKModifierFormat output);
+NSUInteger HKUtilsConvertModifier(NSUInteger modifier, HKModifierFormat input, HKModifierFormat output);
